Create new transaction components for all of the transaction pages
