+++ packages/frontend/containers/incomes/income.add.container.tsx
@@ -1,14 +1,20 @@
-import { useState } from 'react';
+import { useMemo } from 'react';
+import { useDispatch } from 'react-redux';
 
 import {
   CreateIncomeDto,
+  TransactionTypeEnum,
   useIncomesCreateMutation,
   useTransactionTemplatesFindOneQuery,
 } from '$api/generated/financerApi';
 import { DataHandler } from '$blocks/data-handler/data-handler';
+import { ToastMessageTypes } from '$blocks/toast/toast';
+import { TransactionForm } from '$blocks/transaction-form/transaction-form';
+import { TransactionTemplateSwitcher } from '$blocks/transaction-template-switcher/transaction-template-switcher';
 import { useUserDefaultIncomeAccount } from '$hooks/settings/user-preference/useUserDefaultIncomeAccount';
 import { useViewTransitionRouter } from '$hooks/useViewTransitionRouter';
-import { IncomeAdd } from '$pages/incomes/income.add';
+import { addToastMessage } from '$reducer/notifications.reducer';
+import { UpdatePageInfo } from '$renderers/seo/updatePageInfo';
 import { parseErrorMessagesToArray } from '$utils/apiHelper';
 
 interface IncomeAddContainerProps {
@@ -17,10 +23,11 @@ interface IncomeAddContainerProps {
 
 export const IncomeAddContainer = ({ templateId }: IncomeAddContainerProps) => {
   const { push } = useViewTransitionRouter();
-  const [errors, setErrors] = useState<string[]>([]);
-  const [addIncome, { isLoading: isCreating }] = useIncomesCreateMutation();
-  const { data: defaultIncomeAccount, isLoading: isLoadingDefaultAccount } =
-    useUserDefaultIncomeAccount({ skip: !!templateId });
+  const [addIncome] = useIncomesCreateMutation();
+  const { data: defaultIncomeAccount } = useUserDefaultIncomeAccount({
+    skip: !!templateId,
+  });
+  const dispatch = useDispatch();
 
   const templateData = useTransactionTemplatesFindOneQuery(
     { id: templateId as string },
@@ -29,17 +36,23 @@ export const IncomeAddContainer = ({ templateId }: IncomeAddContainerProps) => {
 
   const { data: transactionTemplate } = templateData;
 
-  const handleSubmit = async (newIncomeData: CreateIncomeDto) => {
+  const handleSubmit = async (createIncomeDto: CreateIncomeDto) => {
     try {
-      const { _id: id } = await addIncome({
-        createIncomeDto: newIncomeData,
-      }).unwrap();
+      const { _id: id } = await addIncome({ createIncomeDto }).unwrap();
 
       push(`/statistics/incomes/${id}`);
       // eslint-disable-next-line @typescript-eslint/no-explicit-any
     } catch (error: any) {
       if (error.status === 400 || error.status === 404) {
-        setErrors(parseErrorMessagesToArray(error?.data?.message));
+        dispatch(
+          addToastMessage({
+            type: ToastMessageTypes.ERROR,
+            message: 'Submission failed',
+            additionalInformation: parseErrorMessagesToArray(
+              error?.data?.message
+            ),
+          })
+        );
         return;
       }
 
@@ -48,19 +61,38 @@ export const IncomeAddContainer = ({ templateId }: IncomeAddContainerProps) => {
     }
   };
 
-  const isLoading = isLoadingDefaultAccount;
+  const initialValues = useMemo(() => {
+    if (!transactionTemplate) {
+      return { toAccount: defaultIncomeAccount };
+    }
+    const categories = transactionTemplate?.categories?.map((categoryId) => ({
+      category_id: categoryId,
+      amount: NaN,
+    }));
+
+    return {
+      ...transactionTemplate,
+      categories,
+    };
+  }, [defaultIncomeAccount, transactionTemplate]);
 
   return (
     <>
       <DataHandler skipNotFound {...templateData} />
+      <UpdatePageInfo
+        title="Add Income"
+        headerAction={
+          <TransactionTemplateSwitcher
+            selectedTemplate={templateId}
+            templateType={TransactionTypeEnum.Income}
+          />
+        }
+      />
       {(!templateId || transactionTemplate) && (
-        <IncomeAdd
-          defaultIncomeAccount={defaultIncomeAccount}
-          template={transactionTemplate}
-          isLoading={isLoading}
-          isCreating={isCreating}
-          errors={errors}
+        <TransactionForm
+          initialValues={initialValues}
           onSubmit={handleSubmit}
+          hasToAccountField
         />
       )}
     </>
